Move storage info in its own file
